fix(infer): databricks types corrected against Spark's own analyzer (430 -> 5)
The Spark-goldens gate's debt ratchet drops from 430 wrong columns to 5.
Registry: ceil/floor scale-arg form and decimal input, date_add/dateadd
arity split (2-arg DATE, 3-arg TIMESTAMP), datediff arity split, sign/
signum double, getbit/bit_get tinyint, array_position bigint, bit_count
int, type-named converters (smallint/tinyint), positive/negative/abs
string crosscast, zeroifnull widening, avg/mean decimal-preserving,
uniform NULL-guard, percentile_disc absent by design, try_* interval
rules, date_part/datepart via a new behavior special hook (field-value-
keyed: second family decimal, NULL double, non-literal unknown).
Literals: integral sizing int/bigint/decimal by magnitude, decimal-point
literals are DECIMAL not double, suffixes L/S/Y/BD/F/D, timestamp_ntz/ltz
literal prefixes. Operators (shared, dialect-guarded): div is bigint under
float-division, interval arithmetic stays interval, timestamp±interval
keeps timestamp while date±interval is honestly unknown (flavor-dependent),
date/timestamp/time subtraction yields interval via a new dateSubtraction
behavior knob, float-division keeps decimal for decimal-with-exact
operands. Coercion: approximate types now rank above decimal (every
engine's promotion rule; the old order claimed decimal for 2.35E10 * 1.0).
from_json/from_csv accept DDL table-schema strings as an implicit struct
instead of fabricating a mushed scalar. And one real lowering bug: a
nested/chained cast took the INNER cast's dataType as the outer typeText
(x::long::int read as long) — direct-child capture fixes the class.

The five remaining wrong columns are ledgered in #40 (date_part over an
interval source, the float+decimal cross-dialect promotion conflict,
multi-row VALUES column union).

+++ src/infer/coerce.ts
@@ -5,14 +5,18 @@ import { UNKNOWN, type Type } from "./types.js";
 // don't coerce (→ unknown). `unknown` is contagious for operators but filtered for
 // "common type of a list" (see commonType).
 
+// Approximate types dominate exact ones: DECIMAL widens past the integers but LOSES to
+// FLOAT/DOUBLE (Spark, Postgres, T-SQL, Snowflake all promote decimal-with-approximate to
+// the approximate side). The old order had decimal on top, so `2.35E10 * 1.0` claimed
+// decimal where every engine answers double — caught by the Spark-goldens gate.
 const NUMERIC_RANK: Record<string, number> = {
 	tinyint: 1,
 	smallint: 2,
 	int: 3,
 	bigint: 4,
-	float: 5,
-	double: 6,
-	decimal: 7,
+	decimal: 5,
+	float: 6,
+	double: 7,
 };
 
 /** The wider of two types, or `unknown` when they don't coerce. */
